% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_seq_analysis.R
\name{base.freq}
\alias{base.freq}
\title{Title Calculate the base frequency}
\usage{
base.freq(inputseq)
}
\arguments{
\item{inputseq}{input sequence whose seperation is ''}
}
\value{
a table of each sequence's frequency
}
\description{
Title Calculate the base frequency
}
\examples{
base.freq("AGTCAGTTTAC")
}
