% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_seq_analysis.R
\name{Blast}
\alias{Blast}
\title{Title simplified global alignment of 2 sequences}
\usage{
Blast(x, y)
}
\arguments{
\item{x}{input sequence1 for alignment whose seperation is ''}

\item{y}{input sequence2 for alignment whose seperation is ''}
}
\value{
Dynamic programming matrix,Trace path matrix,Alignment
}
\description{
Title simplified global alignment of 2 sequences
}
\examples{
Blast("agctattgca","agtcattgca"),Blast(seq1,seq2)
}
