% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_seq_analysis.R
\name{download_genbank}
\alias{download_genbank}
\title{Title Batch download genbank/fasta file from URL}
\usage{
download_genbank(acc, file_type, db_type)
}
\arguments{
\item{acc}{accession number or a list of accession number}

\item{file_type}{the type of file you want to download}

\item{db_type}{the database you want to search that help narrow down the range}
}
\value{
specified type of file(fasta or genbank or other)
}
\description{
Title Batch download genbank/fasta file from URL
}
\examples{
download_genbank(accn,"gb","nucleotide")
}
