% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_seq_analysis.R
\name{gb2fas}
\alias{gb2fas}
\title{Title File format conversion:genbank to fasta}
\usage{
gb2fas(path)
}
\arguments{
\item{path}{the path of the required file}
}
\value{
fasta file
}
\description{
Title File format conversion:genbank to fasta
}
