% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_seq_analysis.R
\name{parse_fasta}
\alias{parse_fasta}
\title{Title Parse fasta file(Calculate the base frequency)}
\usage{
parse_fasta(lst)
}
\arguments{
\item{lst}{a list of id and seq}
}
\value{
res :a table of each sequence's frequency
}
\description{
Title Parse fasta file(Calculate the base frequency)
}
\examples{
pparse_fasta(lt)
}
