% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_seq_analysis.R
\name{Blast}
\alias{Blast}
\title{Title Simplified global alignment of 2 sequences}
\usage{
Blast(seq1, seq2, score = list(match = 5, mismatch = -2, GAP = -6))
}
\arguments{
\item{seq1}{input sequence1 for alignment whose seperation is ''(a character string)}

\item{seq2}{input sequence2 for alignment whose seperation is ''(a character string)}

\item{score}{score system}
}
\value{
Align object
}
\description{
Title Simplified global alignment of 2 sequences
}
\examples{
Blast("agctattgca","agtcattgca"),Blast(seq1,seq2)
}
