% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_seq_analysis.R
\name{parse_fasta}
\alias{parse_fasta}
\title{Title Read fasta file and calculate the base frequency}
\usage{
parse_fasta(path)
}
\arguments{
\item{path}{the path of the required file}
}
\value{
a table of each sequence's frequency
}
\description{
Title Read fasta file and calculate the base frequency
}
\examples{
parse_fasta("flu_seq.fas")
}
