% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_seq_analysis.R
\name{read_fasta}
\alias{read_fasta}
\title{Title Read fasta file}
\usage{
read_fasta(path)
}
\arguments{
\item{path}{the path of the required file}
}
\value{
sequence in fasta file and description
}
\description{
Title Read fasta file
}
\examples{
read_fasta("flu_seq.fas")
}
